buffer size to long instead of int
Updated host buffer size to long instead of int.
Added {get,parse}Long methods to Settings

+++ src/routing/MaxPropRouterWithEstimation.java
@@ -516,9 +516,9 @@ public class MaxPropRouterWithEstimation extends ActiveRouter {
 	 */
 	public int calcThreshold() {
 		/* b, x and p refer to respective variables in the paper's equations */
-		int b = this.getBufferSize();
-		int x = this.avgTransferredBytes;
-		int p;
+		long b = this.getBufferSize();
+		long x = this.avgTransferredBytes;
+		long p;
 
 		if (x == 0) {
 			/* can't calc the threshold because there's no transfer data */
